// client/services/socket/socket.events.ts
export enum SocketEventType {
  MESSAGE = "message",

  TYPING = "typing",

  ONLINE_STATUS = "online_status",

  USER_UPDATE = "user_update",

  CHAT_DELETED = "chat_deleted",

  LOCATION = "location",

  MAP_VISIBILITY = "map_visibility",

  MAP_UPDATE = "map_update",

  MESSAGE_READ = "message_read",

  MESSAGE_DELIVERED = "message_delivered",

  MESSAGE_DELETED = "message_deleted",

  MESSAGE_EDITED = "message_edited",

  STORY = "story",

  CONVERSATION_UPDATED = "conversation_updated",

  PING = "ping",

  PONG = "pong",

  MESSAGE_ACK="message_ack"
}